Add Save and Save As distinction
Each Jtool session, for each map to save, the user must complete the Save dialog to give the program permission to write. So the first save will always be a save as, but subsequent ones will be quicksaves. I also made Save As re-prompt when revisiting a map in the same session for a more consistent feeling flow.